+++ Plugin/Thunderstore/CHANGELOG.md
@@ -1,7 +1,17 @@
+## 1.0.2
+A lot of minor fixes
+- Added more layers to the navMesh of the Raptor
+    - It should now be able to reach more places.
+- The Raptor should now no longer attempt to reach/chase players inside the ship if it is not in the ship itself
+- The Raptor should now check if it has a clear path before attempting to pounce on the player
+    - improved pounce animation
+- The Raptor should now correctly begin chasing the player who hit it (during a chase)
+- The Raptor should no longer play the hit sound effect once its dead and gets hit
+
 ## 1.0.1
 Small fix.
 - Fixed the wrong assetbundle being packaged.
-    - This should make the raptor be able to load in game
+    - This should make the Raptor be able to load in game
     - This could also update some of its behaviour as the wrong/old assets were bundled.
 
 ## 1.0.0